Rename services to match convention Refactoring of the compute service
